functions/kucoin_buy/requirements.txt
python-kucoin